django-app/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
django-cors-headers==4.3.0
mysqlclient==2.2.0
django-simple-history==3.4.0
djangorestframework-simplejwt==5.3.0